GA-193 merge request

+++ kairos-gdpr/src/main/java/com/kairos/controller/master_data_management/asset_management/AssetTypeController.java
@@ -38,78 +38,100 @@ public class AssetTypeController {
 
     @ApiOperation("add AssetType")
     @PostMapping("/add")
-    public ResponseEntity<Object> createStorageType(@PathVariable Long countryId,@PathVariable Long organizationId, @Valid @RequestBody ValidateListOfRequestBody<AssetType> assetTypes) {
+    public ResponseEntity<Object> createStorageType(@PathVariable Long countryId, @PathVariable Long organizationId, @Valid @RequestBody ValidateListOfRequestBody<AssetType> assetTypes) {
         if (countryId == null) {
-            return ResponseHandler.invalidResponse(HttpStatus.BAD_REQUEST, false, "country id is null");
+            return ResponseHandler.invalidResponse(HttpStatus.BAD_REQUEST, false, "country id can't be null");
+        } else if (organizationId == null) {
+            return ResponseHandler.invalidResponse(HttpStatus.BAD_REQUEST, false, "organization id can't be null");
         }
-        return ResponseHandler.generateResponse(HttpStatus.OK, true, storageTypeService.createAssetType(countryId,organizationId,assetTypes.getRequestBody()));
+        return ResponseHandler.generateResponse(HttpStatus.OK, true, storageTypeService.createAssetType(countryId, organizationId, assetTypes.getRequestBody()));
 
     }
 
 
     @ApiOperation("get AssetType by id")
     @GetMapping("/{id}")
-    public ResponseEntity<Object> getStorageType(@PathVariable Long countryId,@PathVariable Long organizationId,@PathVariable BigInteger id) {
+    public ResponseEntity<Object> getStorageType(@PathVariable Long countryId, @PathVariable Long organizationId, @PathVariable BigInteger id) {
         if (id == null) {
-            return ResponseHandler.invalidResponse(HttpStatus.BAD_REQUEST, false, "id is null");
-        } else if (countryId == null) {
-            return ResponseHandler.invalidResponse(HttpStatus.BAD_REQUEST, false, "country id is null");
-
-        } else
-            return ResponseHandler.generateResponse(HttpStatus.OK, true, storageTypeService.getAssetType(countryId,organizationId,id));
+            return ResponseHandler.invalidResponse(HttpStatus.BAD_REQUEST, false, "id cannot be null");
+        }
+        if (countryId == null) {
+            return ResponseHandler.invalidResponse(HttpStatus.BAD_REQUEST, false, "country id can't be null");
+        }
+        if (organizationId == null) {
+            return ResponseHandler.invalidResponse(HttpStatus.BAD_REQUEST, false, "organization id can't be null");
+        }
+        return ResponseHandler.generateResponse(HttpStatus.OK, true, storageTypeService.getAssetType(countryId, organizationId, id));
 
     }
 
 
     @ApiOperation("get all AssetType ")
     @GetMapping("/all")
-    public ResponseEntity<Object> getAllStorageType(@PathVariable Long countryId,@PathVariable Long organizationId) {
-        return ResponseHandler.generateResponse(HttpStatus.OK, true, storageTypeService.getAllAssetType(countryId,organizationId));
+    public ResponseEntity<Object> getAllStorageType(@PathVariable Long countryId, @PathVariable Long organizationId) {
+        if (countryId == null) {
+            return ResponseHandler.invalidResponse(HttpStatus.BAD_REQUEST, false, "country id can't be null");
+        }
+        if (organizationId == null) {
+            return ResponseHandler.invalidResponse(HttpStatus.BAD_REQUEST, false, "organization id can't be null");
+        }
+        return ResponseHandler.generateResponse(HttpStatus.OK, true, storageTypeService.getAllAssetType(countryId, organizationId));
 
     }
 
 
     @ApiOperation("get AssetType by name")
     @GetMapping("/name")
-    public ResponseEntity<Object> getStorageTypeByName(@PathVariable Long countryId,@PathVariable Long organizationId, @RequestParam String name) {
-        return ResponseHandler.generateResponse(HttpStatus.OK, true, storageTypeService.getAssetTypeByName(countryId,organizationId,name));
+    public ResponseEntity<Object> getStorageTypeByName(@PathVariable Long countryId, @PathVariable Long organizationId, @RequestParam String name) {
+        return ResponseHandler.generateResponse(HttpStatus.OK, true, storageTypeService.getAssetTypeByName(countryId, organizationId, name));
 
     }
 
 
     @ApiOperation("delete AssetType  by id")
     @DeleteMapping("/delete/{id}")
-    public ResponseEntity<Object> deleteStorageType(@PathVariable Long countryId,@PathVariable Long organizationId,@PathVariable BigInteger id) {
+    public ResponseEntity<Object> deleteStorageType(@PathVariable Long countryId, @PathVariable Long organizationId, @PathVariable BigInteger id) {
         if (id == null) {
-            return ResponseHandler.invalidResponse(HttpStatus.BAD_REQUEST, false, "id is null");
-        } else if (countryId == null) {
-            return ResponseHandler.invalidResponse(HttpStatus.BAD_REQUEST, false, "country id is null");
+            return ResponseHandler.invalidResponse(HttpStatus.BAD_REQUEST, false, "id cannot be null");
         }
-        return ResponseHandler.generateResponse(HttpStatus.OK, true, storageTypeService.deleteAssetType(countryId,organizationId,id));
+        if (countryId == null) {
+            return ResponseHandler.invalidResponse(HttpStatus.BAD_REQUEST, false, "country id can't be null");
+        }
+        if (organizationId == null) {
+            return ResponseHandler.invalidResponse(HttpStatus.BAD_REQUEST, false, "organization id can't be null");
+
+        }
+        return ResponseHandler.generateResponse(HttpStatus.OK, true, storageTypeService.deleteAssetType(countryId, organizationId, id));
 
     }
 
     @ApiOperation("update subAsset by id")
     @PutMapping("/update/{id}")
-    public ResponseEntity<Object> updateStorageType(@PathVariable Long countryId,@PathVariable Long organizationId,@PathVariable BigInteger id,  @Valid @RequestBody AssetType assetType) {
+    public ResponseEntity<Object> updateStorageType(@PathVariable Long countryId, @PathVariable Long organizationId, @PathVariable BigInteger id, @Valid @RequestBody AssetType assetType) {
         if (id == null) {
-            return ResponseHandler.invalidResponse(HttpStatus.BAD_REQUEST, false, "id is null");
+            return ResponseHandler.invalidResponse(HttpStatus.BAD_REQUEST, false, "id cannot be null");
         }
+        if (countryId == null) {
+            return ResponseHandler.invalidResponse(HttpStatus.BAD_REQUEST, false, "country id can't be null");
+        }
+        if (organizationId == null) {
+            return ResponseHandler.invalidResponse(HttpStatus.BAD_REQUEST, false, "organization id can't be null");
 
-        return ResponseHandler.generateResponse(HttpStatus.OK, true, storageTypeService.updateAssetType(countryId,organizationId,id, assetType));
+        }
+        return ResponseHandler.generateResponse(HttpStatus.OK, true, storageTypeService.updateAssetType(countryId, organizationId, id, assetType));
 
     }
 
 
-
     @ApiOperation("update subAsset by id")
     @PostMapping("/subAsset/add")
-    public ResponseEntity<Object> addSubAssetTypes(@PathVariable Long countryId,@PathVariable Long organizationId, @Valid @RequestBody AssetTypeDTO assetTypeDto) {
+    public ResponseEntity<Object> addSubAssetTypes(@PathVariable Long countryId, @PathVariable Long organizationId, @Valid @RequestBody AssetTypeDTO assetTypeDto) {
         if (countryId == null) {
-            return ResponseHandler.invalidResponse(HttpStatus.BAD_REQUEST, false, "id is null");
+            return ResponseHandler.invalidResponse(HttpStatus.BAD_REQUEST, false, "country id can't be null");
+        } else if (organizationId == null) {
+            return ResponseHandler.invalidResponse(HttpStatus.BAD_REQUEST, false, "organization id can't be null");
         }
-
-        return ResponseHandler.generateResponse(HttpStatus.OK, true, storageTypeService.createAssetTypeAndAddSubAssetTypes(countryId,organizationId,assetTypeDto));
+        return ResponseHandler.generateResponse(HttpStatus.OK, true, storageTypeService.createAssetTypeAndAddSubAssetTypes(countryId, organizationId, assetTypeDto));
     }
